agentic/requirements.txt
# =============================================================================
# Python AI agent runtime dependencies
#
# Pruned to only what is actually imported today. When new code lands that
# needs FastAPI, LangGraph, Pydantic, etc., add them back here in the same
# pinned style.
# =============================================================================

# ── Knowledge graph driver ────────────────────────────────────────────────────
neo4j==5.26.0

# ── LLM client (CLI test bot) ────────────────────────────────────────────────
langchain-core==0.3.84
langchain-openai==0.2.14
langchain-anthropic==0.3.7

# ── Test framework ───────────────────────────────────────────────────────────
pytest==8.3.4
pytest-asyncio==0.24.0

asyncpg==0.31.0
